+++ s2rest_play/conf/routes
@@ -106,8 +106,7 @@ POST        /graphs/copyLabel/:oldLabelName/:newLabelName                 contro
 POST        /graphs/renameLabel/:oldLabelName/:newLabelName               controllers.AdminController.renameLabel(oldLabelName, newLabelName)
 POST        /graphs/updateHTable/:labelName/:newHTableName                controllers.AdminController.updateHTable(labelName, newHTableName)
 PUT         /graphs/loadCache                                             controllers.AdminController.loadCache()
-PUT         /graphs/invalidateCache/:labelName                            controllers.AdminController.invalidateCache(labelName)
-
+POST        /graphs/invalidateCache/:labelName                            controllers.AdminController.invalidateCache(labelName)
 
 # Counter Admin API
 POST    /counter/v1/:service/:action                                    controllers.CounterController.createAction(service, action)
